feat: convert workplaces into built-in extension
