+++ app/api/content/workspaces/[id]/state/route.ts
@@ -1,45 +1,1 @@
-import { NextRequest, NextResponse } from "next/server";
-import { requireAuth } from "@/lib/infrastructure/auth/middleware";
-import { saveWorkspaceState } from "@/lib/domain/workspaces";
-
-type Params = Promise<{ id: string }>;
-
-function errorResponse(error: unknown, fallback: string) {
-  const message = error instanceof Error ? error.message : fallback;
-  const isAuthError = message.toLowerCase().includes("auth");
-  return NextResponse.json(
-    {
-      success: false,
-      error: {
-        code: isAuthError ? "UNAUTHORIZED" : "INTERNAL_ERROR",
-        message: isAuthError ? "Authentication required" : fallback,
-      },
-    },
-    { status: isAuthError ? 401 : 500 }
-  );
-}
-
-export async function PATCH(
-  request: NextRequest,
-  { params }: { params: Params }
-) {
-  try {
-    const session = await requireAuth();
-    const { id } = await params;
-    const body = await request.json();
-    const data = await saveWorkspaceState(session.user.id, id, body);
-    if (!data) {
-      return NextResponse.json(
-        {
-          success: false,
-          error: { code: "NOT_FOUND", message: "Workspace not found" },
-        },
-        { status: 404 }
-      );
-    }
-    return NextResponse.json({ success: true, data });
-  } catch (error) {
-    console.error("[Workspace State API] PATCH error:", error);
-    return errorResponse(error, "Failed to save workspace state");
-  }
-}
+export { handleSaveWorkplaceState as PATCH } from "@/extensions/workplaces/server/routes";
