+++ lib/extensions/README.md
@@ -6,14 +6,21 @@ Built-in extensions are first-party feature modules with clear ownership boundar
 - Create `extensions/<name>/`
 - Add the extension manifest, runtime, and server runtime in that folder
 - Register the extension once in `lib/extensions/installed.ts`
+- If the extension should be user-toggleable, set `canDisable: true` in the manifest
+- If the extension needs top-level shell UI, contribute it through the runtime shell slots instead of importing the feature directly into shared shell code
 
 ## Expected Structure
 - `manifest.ts`: metadata, nav items, settings metadata, auth metadata
-- `client.tsx`: client runtime contributions such as surfaces, dialogs, slash commands, editor blocks
+- `client.tsx`: client runtime contributions such as surfaces, shell controls, dialogs, slash commands, editor blocks
 - `server-runtime.ts`: server-safe editor/runtime contributions
 - `components/`: UI owned by the extension
 - `server/`: services, types, parsers, and thin route handlers owned by the extension
 - `state/`: extension-local stores
+
+## Activation Rules
+- Runtime enablement is client-persisted and layered on top of `enabledByDefault`
+- Disabled extensions must disappear through the registry filters, not through direct feature conditionals in shared UI
+- If an extension is disabled, its shell controls, dialogs, settings dialog, and runtime hooks should not mount
 
 ## Compatibility Rule
 Legacy paths outside `extensions/<name>` may remain temporarily, but they should only re-export extension-owned code. New logic should be added inside the extension folder.
